refactor: 팔로우버튼 컴포넌트로 분리
